3/1 empezando

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -44,7 +44,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setFloat(idx, comercial.getComision());
+			ps.setDouble(idx, comercial.getComision());
 			return ps;
 		},keyHolder);
